addressing some comments
