+++ cns/restserver/restserver.go
@@ -59,6 +59,7 @@ type containerstatus struct {
 	VMVersion                     string
 	HostVersion                   string
 	CreateNetworkContainerRequest cns.CreateNetworkContainerRequest
+	WaitingForUpdate              bool // True when NC is waiting for NMA to sync versions/rules
 }
 
 // httpRestServiceState contains the state we would like to persist.
